refactor POC
